lint: `LintDiagnosticBuilder` into `rustc_errors`
